+++ client/src/components/atoms/TransparentButton/index.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import styled from '@emotion/styled';
+
+interface IProps {
+  onClick: (event: React.MouseEvent<HTMLButtonElement>) => void;
+}
+
+const StyledTransparentButton = styled.button`
+  border: none;
+  background: transparent;
+`;
+const TransparentButton: React.FC<IProps> = ({ children, onClick }) => {
+  return <StyledTransparentButton onClick={onClick}>{children}</StyledTransparentButton>;
+};
+
+export default TransparentButton;
