+++ client/src/components/atoms/index.tsx
@@ -1,5 +1,6 @@
 export { default as BoxButton } from 'components/atoms/BoxButton';
 export { default as Input } from 'components/atoms/Input';
-export { default as IconButton } from 'components/atoms/IconButton';
 export { default as Toast } from 'components/atoms/Toast';
 export { default as Title } from 'components/atoms/Title';
+export { default as IconImg } from 'components/atoms/IconImg';
+export { default as TransparentButton } from 'components/atoms/TransparentButton';
